Alteração tipo colunas weight e height

+++ src/database/migrations/20191226180051-create-students.js
@@ -22,12 +22,12 @@ module.exports = {
         allowNull: false,
       },
       weight: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DOUBLE,
         defaultValue: 0.0,
         allowNull: false,
       },
       height: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DOUBLE,
         defaultValue: 0.0,
         allowNull: false,
       },
